/*Update: *Change calculation set_temp *Change coeff. PID: Kp = 0.40, Ki = 0.35, Kd = 0.45 */

+++ Read_Battery.h
@@ -3,6 +3,6 @@
 
 void Read_Input_Volt();
 void Check_Input_Voltage(double check_volt);
-int GetInputVolt();
+double GetInputVolt();
 
 #endif
